apps/ambassador/requirements.txt
Django>=4.1.5<4.2
djangorestframework>=3.14.0<3.15
kafka-python3>=3.0.0<3.1
drf-spectacular>=0.25.1<0.26
psycopg2-binary>=2.9.5<2.10
